More work

+++ server/src/main/java/ch/rasc/ngstart/service/UserController.java
@@ -2,18 +2,24 @@ package ch.rasc.ngstart.service;
 
 import java.util.List;
 
-import org.springframework.http.HttpStatus;
+import javax.validation.Valid;
+
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.querydsl.core.types.dsl.Expressions;
+import com.querydsl.jpa.impl.JPAQuery;
+
 import ch.rasc.ngstart.JPAQueryFactory;
+import ch.rasc.ngstart.dto.CrudDeleteResponse;
+import ch.rasc.ngstart.dto.CrudUpdateResponse;
 import ch.rasc.ngstart.entity.QUser;
 import ch.rasc.ngstart.entity.User;
 
@@ -40,8 +46,23 @@ public class UserController {
 
 	@PostMapping("/be/user-save")
 	@PreAuthorize("hasAuthority('ADMIN')")
-	@ResponseStatus(code = HttpStatus.NO_CONTENT)
-	public void saveUser(@RequestBody User user) {
+	public CrudUpdateResponse saveUser(@RequestBody @Valid User user,
+			BindingResult bindingResult) {
+
+		if (bindingResult.hasErrors()) {
+			return CrudUpdateResponse.error(bindingResult);
+		}
+
+		JPAQuery<Integer> query = this.jpaQueryFactory.selectOne().from(QUser.user)
+				.where(QUser.user.userName.equalsIgnoreCase(user.getUserName()));
+		if (user.getId() != null) {
+			query.where(QUser.user.id.ne(user.getId()));
+		}
+		if (query.fetchCount() > 0) {
+			bindingResult.rejectValue("userName", "Username already taken");
+			return CrudUpdateResponse.error(bindingResult);
+		}
+
 		if (StringUtils.hasText(user.getPasswordHash())) {
 			user.setPasswordHash(this.passwordEncoder.encode(user.getPasswordHash()));
 		}
@@ -50,14 +71,28 @@ public class UserController {
 					.from(QUser.user).where(QUser.user.id.eq(user.getId())).fetchOne());
 		}
 		this.jpaQueryFactory.getEntityManager().merge(user);
+		return CrudUpdateResponse.updatedSuccessful();
 	}
 
 	@PostMapping("/be/user-delete")
 	@PreAuthorize("hasAuthority('ADMIN')")
-	@ResponseStatus(code = HttpStatus.NO_CONTENT)
-	public void deleteUser(@RequestBody long userId) {
-		User user = this.jpaQueryFactory.getEntityManager().find(User.class, userId);
-		this.jpaQueryFactory.getEntityManager().remove(user);
+	public CrudDeleteResponse deleteUser(@RequestBody long userId) {
+		if (!isLastAdmin(userId)) {
+			User user = this.jpaQueryFactory.getEntityManager().find(User.class, userId);
+			this.jpaQueryFactory.getEntityManager().remove(user);
+			return CrudDeleteResponse.success();
+		}
+
+		return CrudDeleteResponse.error("Can't delete last ADMIN user");
+	}
+
+	private boolean isLastAdmin(Long id) {
+		JPAQuery<Integer> query = this.jpaQueryFactory.select(Expressions.ONE)
+				.from(QUser.user)
+				.where(QUser.user.id.ne(id).and(QUser.user.enabled.isTrue())
+						.and(QUser.user.authorities.eq("ADMIN")));
+
+		return query.fetchFirst() == null;
 	}
 
 }
